Add files via upload
This project has two functionalities: to record temperature and to detect LPG (liquified petroleum gas). It regulates the speed of the fan with respect to the recorded temperature. When the sensor detects LPG, a buzzer is set, and the user is alerted that the LPG leak has been detected. This system changes the speed of the fan according to the temperature and is thus helpful in making a power-saving system.